third_party/WebKit/LayoutTests/inspector-protocol/dom/dom-collect-class-names.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

function test()
{
    var rootNode;
    var classNames= [];

    InspectorTest.sendCommand("DOM.getDocument", {}, onGotDocument);
    InspectorTest.sendCommand("DOM.enable", {});
    InspectorTest.eventHandler["DOM.setChildNodes"] = setChildNodes;

    function onGotDocument(msg)
    {
        rootNode = msg.result.root;
        InspectorTest.sendCommandOrDie("DOM.collectClassNamesFromSubtree", { nodeId: rootNode.nodeId }, onClassNamesCollected);
    }

    function collectClassNamesFromSubtree()
    {
        InspectorTest.sendCommand("DOM.requestChildNodes", { nodeId: rootNode.children[0].children[1].nodeId }, null);
    }

    function setChildNodes(response)
    {
        var nodes = response.params.nodes;
        InspectorTest.sendCommandOrDie("DOM.collectClassNamesFromSubtree", { nodeId: nodes[1].nodeId }, onSubtreeClassNamesCollected);
    }

    function onSubtreeClassNamesCollected(response)
    {
        var subtreeClassNames = response.classNames.sort();
        InspectorTest.log("All class names: ");
        for (var i = 0; i < classNames.length; i++)
            InspectorTest.log(classNames[i]);
        InspectorTest.log("Subtree class names: ");
        for (var i = 0; i < subtreeClassNames.length; i++)
            InspectorTest.log(subtreeClassNames[i]);

        InspectorTest.completeTest();
    }

    function onClassNamesCollected(response)
    {
        classNames = response.classNames.sort();
        collectClassNamesFromSubtree();
    }
}

</script>
</head>
<body class="body-class">
<div class="class1"></div>
<div class="class2">
    <ul class="class3">
        <li class="class4"></li>
    </ul>
</div>
<div class="class5 class6"></div>
<div id="shadow-host"></div>
<script type="text/javascript">
    var host = document.querySelector("#shadow-host");
    var root = host.createShadowRoot();
    root.innerHTML = "<div class=\"shadow-class\"></div>";
    runTest();
</script>

</body>
</html>
