third_party/blink/web_tests/inspector-protocol/dom/dom-collect-class-names-expected.txt

Tests collecting class names in DOM domain.
All class names: 
body-class
class1
class2
class3
class4
class5
class6
shadow-class
Subtree class names: 
class2
class3
class4
